Move template rendering to utiliy function
Models should be reserved for things that are saved into the database.  Since views aren't something that are saved, they should be organized some other way.  I realized that having template rendering be part of a "utils" module made more sense, and moved the decorators module to be a file within that, which makes a lot more sense organizationally